# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "ShulNET NodeJS API",
    "description": "API underpinning ShulNET or custom integrations."
  },
  "host": "localhost:6868",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format Bearer &lt;token&rt;"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/account/authenticate": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "schema": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        }
      }
    },
    "/account/refresh-token": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "cookie",
            "description": "The refresh token cookie",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85dcd5e4ad6e3f98607550"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        }
      }
    },
    "/account/revoke-token": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/account/register": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Registration definition.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Registration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/account/": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Account"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/account/{id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
    "Registration": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "brian.kresge@gmail.com"
        },
        "title": {
          "type": "string",
          "example": "None"
        },
        "firstName": {
          "type": "string",
          "example": "Brian"
        },
        "lastName": {
          "type": "string",
          "example": "Kresge"
        },
        "password": {
          "type": "string",
          "example": ""
        },
        "confirmPassword": {
          "type": "string",
          "example": ""
        },
        "acceptTerms": {
          "type": "boolean",
          "example": true
        },
        "role": {
          "type": "string",
          "example": "Admin"
        },
        "created": {
          "type": "string",
          "example": "2021-11-27"
        }
      }
    },
    "Account": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "brian.kresge@gmail.com"
        },
        "passwordHash": {
          "type": "string",
          "example": ""
        },
        "title": {
          "type": "string",
          "example": "None"
        },
        "firstName": {
          "type": "string",
          "example": "Brian"
        },
        "lastName": {
          "type": "string",
          "example": "Kresge"
        },
        "acceptTerms": {
          "type": "boolean",
          "example": true
        },
        "role": {
          "type": "string",
          "example": "Admin"
        },
        "verificationToken": {
          "type": "string",
          "example": ""
        },
        "verified": {
          "type": "string",
          "example": "2021-01-01"
        },
        "resetToken": {
          "type": "object",
          "properties": {
            "token": {
              "type": "string",
              "example": ""
            },
            "expires": {
              "type": "string",
              "example": "2021-01-01"
            }
          }
        },
        "passwordReset": {
          "type": "string",
          "example": "2022-01-01"
        },
        "created": {
          "type": "string",
          "example": "2021-01-01"
        },
        "updated": {
          "type": "string",
          "example": "2021-01-01"
        }
      }
    },
...
  }
}